Rename usecase 11-15 tests to match other tests
