+++ src/UnitTests/Features/Event/UC12-RemoveGuest/RemoveGuestHandlerTests.cs
@@ -4,9 +4,9 @@ using Tests.Common.Factories;
 using Tests.Fakes;
 using VIAEventAssociation.Core.Tools.OperationResult.Errors;
 
-namespace Tests.Features.Event.UC11_AddGuest;
+namespace Tests.Features.Event.UC12_RemoveGuest;
 
-public class AddGuestHandlerTest
+public class RemoveGuestHandlerTests
 {
     private FakeEventRepository EventRepository { get; } = new();
     private FakeUserRepository UserRepository { get; } = new();
@@ -14,15 +14,16 @@ public class AddGuestHandlerTest
     
     // # S1
     [Test]
-    public async Task Create_AddGuestHandler_Success()
+    public async Task Create_RemoveGuestHandler_Success()
     {
         // Arrange
         var @event = EventTestDataFactory.ActivePublicEvent();
-        await EventRepository.AddAsync(@event);
         var user = UserRepository.Users.First();
+        @event.AddGuest(user.Id);    
+        await EventRepository.AddAsync(@event);
 
-        var command = AddGuestCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
-        var handler = new AddGuestHandler(EventRepository, Uow);
+        var command = RemoveGuestCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
+        var handler = new RemoveGuestHandler(EventRepository, Uow);
         
         // Act
         var result = await handler.HandleAsync(command);
@@ -32,22 +33,22 @@ public class AddGuestHandlerTest
         Assert.Multiple(() =>
         {
             Assert.That(result.IsFailure, Is.False);
-            Assert.That(updatedEvent.Participants.Any(id => id.Value == user.Id.Value), Is.True);
+            Assert.That(updatedEvent.Participants.Any(id => id.Value == user.Id.Value), Is.False);
         });
     }
     
     // # F1
     [Test]
-    public async Task Create_AddGuestHandler_Fail()
+    public async Task Create_RemoveGuestHandler_Fail()
     {
         // Arrange
-        var @event = EventTestDataFactory.PublicEvent();
+        var @event = EventTestDataFactory.ActivePublicEventWithGuestAndStartTimeInPast();
             
         await EventRepository.AddAsync(@event);
-        var user = UserRepository.Users.First();
+        var user = @event.Participants.First();
 
-        var command = AddGuestCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
-        var handler = new AddGuestHandler(EventRepository, Uow);
+        var command = RemoveGuestCommand.Create(@event.Id.Value.ToString(), user.Value.ToString());
+        var handler = new RemoveGuestHandler(EventRepository, Uow);
         
         // Act
         var result = await handler.HandleAsync(command);
@@ -58,9 +59,8 @@ public class AddGuestHandlerTest
         Assert.Multiple(() =>
         {
             Assert.That(result.IsFailure, Is.True);
-            Assert.That(updatedEvent.Participants.Any(id => id.Value == user.Id.Value), Is.False);
-            Assert.That(result.Errors.Any(x => x.Code == ErrorCode.RequestToEventThatIsNotActive), Is.True);
-
+            Assert.That(updatedEvent.Participants.Any(id => id.Value == user.Value), Is.True);
+            Assert.That(result.Errors.Any(x => x.Code == ErrorCode.CancelParticipationToEventInThePast), Is.True);
         });
     }
 }
