+++ src/UnitTests/Features/Event/UC13-InviteGuest/InviteGuestCommandTests.cs
@@ -1,23 +1,23 @@
 ﻿using Application.AppEntry.Commands.EventCommands;
 using Tests.Fakes;
 
-namespace Tests.Features.Event.UC14_AcceptInvitation;
+namespace Tests.Features.Event.UC13_InviteGuest;
 
-public class AcceptInvitationCommandTest
+public class InviteGuestCommandTests
 {
     private FakeEventRepository EventRepository { get; } = new();   
     private FakeUserRepository UserRepository { get; } = new();
     
     // # S1
     [Test]
-    public void Create_AcceptInvitationCommand_Success()
+    public void Create_InviteGuestCommand_Success()
     {
         // Arrange
         var @event = EventRepository.Events.First();
         var user = UserRepository.Users.First();
 
         // Act
-        var command = AcceptInvitationCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
+        var command = InviteGuestCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
       
         // Assert
         Assert.Multiple(() =>
