+++ src/UnitTests/Features/Event/UC14-AcceptInvitation/AcceptInvitationCommandTests.cs
@@ -1,23 +1,23 @@
 ﻿using Application.AppEntry.Commands.EventCommands;
 using Tests.Fakes;
 
-namespace Tests.Features.Event.UC15_DeclineInvitation;
+namespace Tests.Features.Event.UC14_AcceptInvitation;
 
-public class DeclineInvitationCommandTest
+public class AcceptInvitationCommandTests
 {
     private FakeEventRepository EventRepository { get; } = new();   
     private FakeUserRepository UserRepository { get; } = new();
     
     // # S1
     [Test]
-    public void Create_DeclineInvitationCommand_Success()
+    public void Create_AcceptInvitationCommand_Success()
     {
         // Arrange
         var @event = EventRepository.Events.First();
         var user = UserRepository.Users.First();
 
         // Act
-        var command = DeclineInvitationCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
+        var command = AcceptInvitationCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
       
         // Assert
         Assert.Multiple(() =>
