+++ src/UnitTests/Features/Event/UC14-AcceptInvitation/AcceptInvitationHandlerTests.cs
@@ -5,9 +5,9 @@ using Tests.Fakes;
 using VIAEventAssociation.Core.Domain.Aggregates.Event.Entities.Invitation.Values;
 using VIAEventAssociation.Core.Tools.OperationResult.Errors;
 
-namespace Tests.Features.Event.UC15_DeclineInvitation;
+namespace Tests.Features.Event.UC14_AcceptInvitation;
 
-public class DeclineInvitationHandlerTest
+public class AcceptInvitationHandlerTests
 {
     private FakeEventRepository EventRepository { get; } = new();
     private FakeUserRepository UserRepository { get; } = new();
@@ -15,15 +15,15 @@ public class DeclineInvitationHandlerTest
     
     // # S1
     [Test]
-    public async Task Create_DeclineInvitationHandler_Success()
+    public async Task Create_AcceptInvitationHandler_Success()
     {
         // Arrange
         var @event = EventTestDataFactory.ActivePublicEventWithPendingInvitation();
         await EventRepository.AddAsync(@event);
         var invitation = @event.Invitations.First();
 
-        var command = DeclineInvitationCommand.Create(@event.Id.Value.ToString(), invitation.GuestId.Value.ToString());
-        var handler = new DeclineInvitationHandler(EventRepository, Uow);
+        var command = AcceptInvitationCommand.Create(@event.Id.Value.ToString(), invitation.GuestId.Value.ToString());
+        var handler = new AcceptInvitationHandler(EventRepository, Uow);
         
         // Act
         var result = await handler.HandleAsync(command);
@@ -35,21 +35,21 @@ public class DeclineInvitationHandlerTest
             var updatedInvitation = updatedEvent.Invitations.FirstOrDefault(x => x.Id == invitation.Id);
             
             Assert.That(result.IsFailure, Is.False);
-            Assert.That(updatedInvitation.Status == InvitationStatus.Rejected, Is.True);
+            Assert.That(updatedInvitation.Status == InvitationStatus.Accepted, Is.True);
         });
     }
     
     // # F1
     [Test]
-    public async Task Create_DeclineInvitationHandler_Fail()
+    public async Task Create_AcceptInvitationHandler_Fail()
     {
         // Arrange
         var @event = EventTestDataFactory.ActivePublicEvent();
         await EventRepository.AddAsync(@event);
         var user = UserRepository.Users.First();
 
-        var command = DeclineInvitationCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
-        var handler = new DeclineInvitationHandler(EventRepository, Uow);
+        var command = AcceptInvitationCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
+        var handler = new AcceptInvitationHandler(EventRepository, Uow);
         
         // Act
         var result = await handler.HandleAsync(command);
@@ -58,7 +58,7 @@ public class DeclineInvitationHandlerTest
         Assert.Multiple(() =>
         {
             Assert.That(result.IsFailure, Is.True);
-            Assert.That(result.Errors.Any(x => x.Code == ErrorCode.InvitationDeclineToGuestNotInvited), Is.True);
+            Assert.That(result.Errors.Any(x => x.Code == ErrorCode.InvitationAcceptToGuestNotInvited), Is.True);
 
         });
     }
