+++ src/UnitTests/Features/Event/UC15-DeclineInvitation/DeclineInvitationCommandTests.cs
@@ -1,23 +1,23 @@
 ﻿using Application.AppEntry.Commands.EventCommands;
 using Tests.Fakes;
 
-namespace Tests.Features.Event.UC13_InviteGuest;
+namespace Tests.Features.Event.UC15_DeclineInvitation;
 
-public class InviteGuestCommandTest
+public class DeclineInvitationCommandTests
 {
     private FakeEventRepository EventRepository { get; } = new();   
     private FakeUserRepository UserRepository { get; } = new();
     
     // # S1
     [Test]
-    public void Create_InviteGuestCommand_Success()
+    public void Create_DeclineInvitationCommand_Success()
     {
         // Arrange
         var @event = EventRepository.Events.First();
         var user = UserRepository.Users.First();
 
         // Act
-        var command = InviteGuestCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
+        var command = DeclineInvitationCommand.Create(@event.Id.Value.ToString(), user.Id.Value.ToString());
       
         // Assert
         Assert.Multiple(() =>
